+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/BlockCompiler.java
@@ -3,9 +3,9 @@ package com.aws.jverify.verifier.compiler;
 import com.aws.jverify.Pure;
 import com.aws.jverify.common.Common;
 import com.aws.jverify.generated.*;
-import com.aws.jverify.verifier.compiler.transformations.DoWhileLoopCompiler;
-import com.aws.jverify.verifier.compiler.transformations.ForLoopCompiler;
-import com.aws.jverify.verifier.compiler.transformations.ImpureStatementExpressionCompiler;
+import com.aws.jverify.verifier.compiler.simplifications.DoWhileLoopCompiler;
+import com.aws.jverify.verifier.compiler.simplifications.ForLoopCompiler;
+import com.aws.jverify.verifier.compiler.simplifications.ImpureStatementExpressionCompiler;
 import com.sun.tools.javac.code.Symbol;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.tree.TreeInfo;
