+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/ClassCompiler.java
@@ -5,10 +5,10 @@ import com.aws.jverify.InheritContract;
 import com.aws.jverify.Modifiable;
 import com.aws.jverify.Pure;
 import com.aws.jverify.generated.*;
-import com.aws.jverify.verifier.compiler.transformations.ExternalContractCompiler;
-import com.aws.jverify.verifier.compiler.transformations.VerifyAnnotationCompiler;
-import com.aws.jverify.verifier.compiler.transformations.temporary.ClassesExtendingClassesCompiler;
-import com.aws.jverify.verifier.compiler.transformations.temporary.RecordCompiler;
+import com.aws.jverify.verifier.compiler.simplifications.ExternalContractCompiler;
+import com.aws.jverify.verifier.compiler.simplifications.VerifyAnnotationCompiler;
+import com.aws.jverify.verifier.compiler.simplifications.temporary.ClassesExtendingClassesCompiler;
+import com.aws.jverify.verifier.compiler.simplifications.temporary.RecordCompiler;
 import com.sun.source.tree.Tree;
 import com.sun.tools.javac.api.JavacTrees;
 import com.sun.tools.javac.code.Flags;
