+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/NameCompiler.java
@@ -1,4 +1,4 @@
-package com.aws.jverify.verifier.compiler.transformations;
+package com.aws.jverify.verifier.compiler.simplifications;
 
 import com.sun.tools.javac.code.Symbol;
